refactor: add log module func for member leaves
